[Feat] #185 - 컴포넌트 분리
날짜, 시간, 장소
